refactor(transitions): extract the transition actions into a mixin
